requirements.txt
git+https://github.com/Pycord-Development/pycord
requests
python-dotenv